Fill cubes

+++ guess-game-server/src/main/java/guess/domain/statistics/olap/dimension/YearDimension.java
@@ -5,8 +5,8 @@ import guess.domain.statistics.olap.DimensionType;
 /**
  * Year dimension.
  */
-public class YearDimension extends Dimension<Long> {
-    public YearDimension(Long value) {
+public class YearDimension extends Dimension<Integer> {
+    public YearDimension(Integer value) {
         super(DimensionType.YEAR, value);
     }
 }
